adds functionality to delete card, removes unnecessary imports
