+++ src/App.js
@@ -44,8 +44,9 @@ function App() {
     }
   }
 
-  //this updates the cards whenever the current board is changed
-  useEffect(() => {
+  // this is used (1) to generate cards on the board whenever there's a change to the current board (currentBoard state),
+  // or, (2) to re-render the cards when a card's like count increases, or (3) to re-render the cards when a card is deleted
+  const renderCards = () => {
     axios.get(`${process.env.REACT_APP_BACKEND_URL}/boards/${currentBoard.board_id}/cards`,
       {
         params: {
@@ -60,8 +61,47 @@ function App() {
         console.log('error in getting card list');
         console.log(error.response)
     });
+  }
+
+  //this updates the cards whenever the current board is changed. This is necessary for CardList to work.
+  useEffect(() => {
+    renderCards();
   }, [currentBoard])
 
+
+  //this increases a card's like count and re-renders all displayed cards
+  const increaseLikeCount = (card_id) => {
+    axios.patch(`${process.env.REACT_APP_BACKEND_URL}/cards/${card_id}/like`,
+    {
+        params: {
+            format: 'json'
+        }
+    })
+    .then( (response) => {
+        renderCards();
+    })
+    .catch( (error) => {
+        console.log(error.response);
+        alert("Could not like card")
+    });
+  }
+
+  //this deletes a card and re-renders all displayed cards to reflect the change
+  const deleteCard = (card_id) => {
+    axios.delete(`${process.env.REACT_APP_BACKEND_URL}/cards/${card_id}`,
+    {
+      params: {
+        format: 'json'
+      }
+    })
+    .then( (response) => {
+      renderCards();
+    })
+    .catch( (error) => {
+      console.log(error.response);
+      alert("Could not delete card")
+    });
+  }
 
   const handleChange = (event) => { 
     selectBoard(event);
@@ -120,7 +160,7 @@ function App() {
         <h3>Create a New Board:</h3>
         <NewBoardForm onSubmitCallback={handleSubmit}></NewBoardForm>
 
-        <Board data={currentBoard} cards={cards}></Board>
+        <Board data={currentBoard} cards={cards} onLikeClickCallback={increaseLikeCount} onDeleteClickCallback={deleteCard}></Board>
 
         <h3>Create a New Card</h3>
         <NewCardForm board_id={currentBoard.board_id}></NewCardForm>
